+++ DFC.App.CareerPath.DraftSegmentService.UnitTests/DraftSegmentServiceTests/DraftSegmentServiceGetAllTests.cs
@@ -2,6 +2,7 @@ using Xunit;
 
 namespace DFC.App.CareerPath.DraftSegmentService.UnitTests
 {
+    [Trait("Draft Segment Service", "GetAll Tests")]
     public class DraftSegmentServiceGetAllTests
     {
         [Fact]
