--- DFC.App.CareerPath.DraftSegmentService.UnitTests/UnitTest1.cs
@@ -1,13 +1,0 @@
-using System;
-using Xunit;
-
-namespace DFC.App.CareerPath.DraftSegmentService.UnitTests
-{
-    public class UnitTest1
-    {
-        [Fact]
-        public void Test1()
-        {
-        }
-    }
-}
