+++ DFC.App.CareerPath.SegmentService.UnitTests/SegmentServiceTests/SegmentServiceGetAllTests.cs
@@ -2,6 +2,7 @@ using Xunit;
 
 namespace DFC.App.CareerPath.SegmentService.UnitTests.SegmentServiceTests
 {
+    [Trait("Segment Service", "GetAll Tests")]
     public class SegmentServiceGetAllTests
     {
         [Fact]
