+++ DFC.App.CareerPath.UnitTests/ControllerTests/SegmentControllerTests/SegmentControllerDocumentTests.cs
@@ -1,0 +1,13 @@
+using Xunit;
+
+namespace DFC.App.CareerPath.UnitTests.SegmentControllerTests
+{
+    [Trait("Segment Controller", "Document Tests")]
+    public class SegmentControllerDocumentTests
+    {
+        [Fact]
+        public void Test1()
+        {
+        }
+    }
+}
