+++ DFC.App.CareerPath.UnitTests/ControllerTests/SegmentControllerTests/SegmentControllerIndexTests.cs
@@ -1,0 +1,13 @@
+using Xunit;
+
+namespace DFC.App.CareerPath.UnitTests.SegmentControllerTests
+{
+    [Trait("Segment Controller", "Index Tests")]
+    public class SegmentControllerIndexTests
+    {
+        [Fact]
+        public void Test1()
+        {
+        }
+    }
+}
